Create movie_API
I wanna show the information about movies to users, so I got a movie API from one site.
And then I save the address into the variable which is called "API_KEY". (I can name this anything like "MOVIE_API", "MOVIE_KEY", etc..)
And I also use useEffect which I use from now on.
In the useEffect, I can call fetch() method here, and then I got a response, and then make it to JSON(), and also add the modifier function which makes the state to false.
And in the HTML section, I put a photo of movies by using <img> tag, and then I make <div> tag one more time, and the <div> tag have to get a key.
So I put a key in the <div> tag, and it's from API data of course.
And in the <div> tag, we can put any tags in here.
So I put a <img> tag, and also <h2> for title, and <p> for summary.
And also put a genre of a movie.
Especially, genre is made of an array, so I put an (gr) which means the item of an array, and (index) for unique character for arguements.
When I make these tags, I'm so confused of syntax.
Those things > {<div><ul>(<li" " >)} blah blah...
But I learned many things whild doing this step.
Thank you.